split docker file in backend to two for each main file
Refactor global.env Adds three new methods, two for initializing env variables for reception api and rest api, third one for checking the variables and give default values or stop the program if the variable is required (db_user, db_password, sbs_source)

+++ backend/internal/sbs/sbsProcessing.go
@@ -29,7 +29,7 @@ func ProcessSbsStream() ([]models.AircraftCurrentModel, error) {
 
 	timer := time.Now()
 	for scanner.Scan() {
-		if diff := time.Since(timer).Seconds(); diff > global.WaitingTime {
+		if diff := time.Since(timer).Seconds(); diff > float64(global.WaitingTime) {
 			break
 		}
